refactor(capabilities): migrate the render caps to the struct-hosted actor macro
Move the RenderCapability and HeadlessRenderCapability NativeActor impls
into their sibling runtime modules under #[runtime] and host #[actor] on
each struct (ADR-0123). Refs #2381.
